[tests] Add common local socket server needed for additional LXD tests
